Move timestamp tests to test files that match project source tree
