--- tests/timestamps/__init__.py
@@ -1,1 +1,0 @@
-# Package initialization for timestamps tests.
